TM-5848, Rework Bid Audit Ui
