+++ src/Components/AdministratorPage/AdministratorPage.jsx
@@ -11,6 +11,8 @@ import FeatureFlags from './FeatureFlags';
 import PanelAdmin from './PanelAdmin';
 import ManageBidSeasons from './ManageBidSeasons/ManageBidSeasons';
 import BidAudit from './bidAudit/BidAudit';
+import BidAuditCategory from './bidAudit/BidAuditCategory';
+import BidAuditGrade from './bidAudit/BidAuditGrade';
 import CycleJobCategories from './CycleJobCategories/CycleJobCategories';
 import ManageEntryLevel from './ManageEntryLevel/ManageEntryLevel';
 import JobCategories from './JobCategories/JobCategories';
@@ -66,7 +68,9 @@ const AdministratorPage = (props) => {
         <Route path="/profile/administrator/featureflags" render={() => <FeatureFlags {...featureFlagsProps} />} />
         <Route path="/profile/administrator/cycles" render={() => <BidCycles />} />
         <Route path="/profile/administrator/managebidseasons" render={() => <ManageBidSeasons />} />
-        <Route path="/profile/administrator/bidaudit" render={() => <BidAudit />} />
+        <Route path="/profile/administrator/bidaudit/category/:id" render={() => <BidAuditCategory />} />
+        <Route path="/profile/administrator/bidaudit/grade/:id" render={() => <BidAuditGrade />} />
+        <Route path="/profile/administrator/bidaudit/" render={() => <BidAudit />} />
         <Route path="/profile/administrator/panel/:pmSeqNum" render={() => <PanelAdmin />} />
         <Route path="/profile/administrator/panel/" render={() => <PanelAdmin />} />
         <Route path="/profile/administrator/cyclejobcategories/" render={() => <CycleJobCategories />} />
